refactor: use existing code directly
